Neatend up method structure of Server Controller class, less repeated code. Still a bunch of cleaning up to do removing the old stuff and fully implementing the rest.
Removed the Pending Delivery from Client as it is now handled by the Delivery Service.
Moved ClientController out of the Network folder.
Added ability for ClientController to actually kind of read and send stuffs. More to follow.
Updated NetCodes with actually used Client stuff and also updated it to reflect the changes in the Controller.
Added support to Country and DataObject for converting the objects to a network string.
Adding a base implementation of ClientState, at the moment just removes functionality from things the Client doesn't need to concern itself with.